Move stencil tests to a separate file
